requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
cryptography==1.8.1
cycler==0.10.0
dateutils==0.6.6
enum34==1.1.6
ftputil==3.3.1
functools32==3.2.3.post2
gsw==3.0.3
idna==2.5
ipaddress==1.0.18
isodate==0.5.4
lxml==3.8.0
matplotlib==2.0.0
netCDF4==1.2.7
nose==1.3.7
numpy==1.12.0
packaging==16.8
pycparser==2.17
pyOpenSSL==16.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
scipy==0.19.0
Shapely==1.5.17.post1
six==1.10.0
subprocess32==3.2.7
